UI, tile and parent grouping
Made the background actually LOOK like a UI and made the information box modular. Will display ability information if it has the stats for it.

Made a tile set and made the background of the game Pretty :3 Will look into making the Bottom UI elements also good looking but indecisive.

Grouped a lot of enemy and player actions together into parents which will allow them to be modular and easily expandable