Inventory, Item, Settings saving and shop
Made the inventory 10 separate objects that fill in one by one until they have max. Made the inventory able to be hovered over to reread what the items do.

Added the wet blade and made it function.

Added an INI file that saves settings for player and abilities.

Added a heal button to the shop.