New items, Timer, sprite updates and sfx
Added 5 new items

Added a new enemy

Introduced the timer into the game. Currently generates mana based on a percentage of the timer remaining at end of turn

Updated LOTS of sprites for the UI
Rearranged the UI to be more pretty and move obvious about the information it's telling

Updated the main menu with a background

Added some SFX for shop and incorrect options. Added new main menu and battle themes.